requirements.txt
pymodbus>=3.7.4
paho-mqtt